Pay no attention to that man behind the curtain
Just tinkering and fiddling and puttering about.  Feel a need for a pause,
review, before push to do a serious round of testing and media updates
(new hard disk images for PERQ-2 configs, extensive round of testing and
a refresh of the User Guide, etc.)

* OptionBoard.cs: Some largely cosmetic updates, update to be more
  consistent with the IOBoard class, thinking ahead to tackling the 3Mbit Ethernet :-)

* OIO.cs:
* PERQLink.cs: Update some neglected older code; remove the Clock()
  call that isn't being used and other minor cleanups. The Link thing is really
  cool and might have other uses if it could be made cross-platform... the
  creaky hamster wheel turns in my rusty brain...

* IOBoard.cs: Housekeeping. Cosmetic cleanups. Maybe fix a teeny tiny
  off-by-1 bug?

* DMARegisters.cs: Move this into its own file, since it's a decent
  sized wad o' code

* PERQemu.csproj: Add the new DMA class file

* IOB.cs:
* CIO.cs:
* EIO.cs:
* IOBus.cs:
* Am9519.cs:
* PERQDMA.cs:
* Z80IOBus.cs:
* VideoController.cs:
* Ether3MbitController.cs:
* MicropolisDiskController.cs: Update the mish-mash of "unhandled
  read/write" messages to use the exception that was defined for that
  and wrap up the IOBus/Z80 IOBus read and write calls to catch and
  log them consistently. It's been there forever, so may as well use it :-)

* NullEthernet.cs:
* Ether10MbitController.cs: Update exceptions and logging, minor
  housekeeping
