+++ PERQemu/Emulator/IO/IOB.cs
@@ -73,8 +73,7 @@ namespace PERQemu.IO
                     return _z80System.ReadData();
 
                 default:
-                    Log.Warn(Category.IO, "Unhandled IOB Read from port {0:x2}", port);
-                    return 0xffff;
+                    throw new UnhandledIORequestException(port);
             }
         }
 
@@ -131,8 +130,7 @@ namespace PERQemu.IO
                     break;
 
                 default:
-                    Log.Warn(Category.IO, "Unhandled IOB Write to port {0:x2}, data {1:x4}", port, value);
-                    break;
+                    throw new UnhandledIORequestException(port, value);
             }
         }
