+++ PERQemu/PERQemu.csproj
@@ -221,6 +221,7 @@
     <Compile Include="UI\Output\TIFFFormatter.cs" />
     <Compile Include="Emulator\IO\DiskDevices\MFMDiskController.cs" />
     <Compile Include="Emulator\IO\Z80\PERQDMA.cs" />
+    <Compile Include="Emulator\IO\DMARegisters.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="COPYING.txt" />
